[TRL-455] chore: 여행 기간 수정 관련 예외들 common 패키지로 이동
