+++ src/main/java/com/cosain/trilo/common/exception/trip/InvalidPeriodException.java
@@ -1,25 +1,25 @@
-package com.cosain.trilo.trip.domain.exception;
+package com.cosain.trilo.common.exception.trip;
 
 import com.cosain.trilo.common.exception.CustomException;
 import org.springframework.http.HttpStatus;
 
-public class EmptyPeriodUpdateException extends CustomException {
+public class InvalidPeriodException extends CustomException {
 
-    private static final String ERROR_CODE = "trip-0006";
+    private static final String ERROR_CODE = "trip-0005";
     private static final HttpStatus HTTP_STATUS = HttpStatus.BAD_REQUEST;
 
-    public EmptyPeriodUpdateException() {
+    public InvalidPeriodException() {
     }
 
-    public EmptyPeriodUpdateException(String debugMessage) {
+    public InvalidPeriodException(String debugMessage) {
         super(debugMessage);
     }
 
-    public EmptyPeriodUpdateException(Throwable cause) {
+    public InvalidPeriodException(Throwable cause) {
         super(cause);
     }
 
-    public EmptyPeriodUpdateException(String debugMessage, Throwable cause) {
+    public InvalidPeriodException(String debugMessage, Throwable cause) {
         super(debugMessage, cause);
     }
