+++ src/main/java/com/cosain/trilo/common/exception/trip/TooLongPeriodException.java
@@ -1,25 +1,25 @@
-package com.cosain.trilo.trip.domain.exception;
+package com.cosain.trilo.common.exception.trip;
 
 import com.cosain.trilo.common.exception.CustomException;
 import org.springframework.http.HttpStatus;
 
-public class InvalidPeriodException extends CustomException {
+public class TooLongPeriodException extends CustomException {
 
-    private static final String ERROR_CODE = "trip-0005";
+    private static final String ERROR_CODE = "trip-0009";
     private static final HttpStatus HTTP_STATUS = HttpStatus.BAD_REQUEST;
 
-    public InvalidPeriodException() {
+    public TooLongPeriodException() {
     }
 
-    public InvalidPeriodException(String debugMessage) {
+    public TooLongPeriodException(String debugMessage) {
         super(debugMessage);
     }
 
-    public InvalidPeriodException(Throwable cause) {
+    public TooLongPeriodException(Throwable cause) {
         super(cause);
     }
 
-    public InvalidPeriodException(String debugMessage, Throwable cause) {
+    public TooLongPeriodException(String debugMessage, Throwable cause) {
         super(debugMessage, cause);
     }
