+++ src/main/java/com/cosain/trilo/trip/application/trip/service/trip_title_update/TripTitleUpdateService.java
@@ -1,6 +1,6 @@
 package com.cosain.trilo.trip.application.trip.service.trip_title_update;
 
-import com.cosain.trilo.trip.application.exception.NoTripUpdateAuthorityException;
+import com.cosain.trilo.common.exception.trip.NoTripUpdateAuthorityException;
 import com.cosain.trilo.common.exception.trip.TripNotFoundException;
 import com.cosain.trilo.trip.domain.entity.Trip;
 import com.cosain.trilo.trip.domain.repository.TripRepository;
