+++ src/main/java/com/cosain/trilo/trip/presentation/trip/TripPeriodUpdateController.java
@@ -5,10 +5,10 @@ import com.cosain.trilo.auth.presentation.Login;
 import com.cosain.trilo.auth.presentation.LoginUser;
 import com.cosain.trilo.common.exception.CustomValidationException;
 import com.cosain.trilo.common.exception.trip.TripNotFoundException;
-import com.cosain.trilo.trip.application.exception.NoTripUpdateAuthorityException;
+import com.cosain.trilo.common.exception.trip.NoTripUpdateAuthorityException;
 import com.cosain.trilo.trip.application.trip.service.trip_period_update.TripPeriodUpdateCommand;
 import com.cosain.trilo.trip.application.trip.service.trip_period_update.TripPeriodUpdateService;
-import com.cosain.trilo.trip.domain.exception.EmptyPeriodUpdateException;
+import com.cosain.trilo.common.exception.trip.EmptyPeriodUpdateException;
 import com.cosain.trilo.trip.presentation.trip.dto.request.TripPeriodUpdateRequest;
 import com.cosain.trilo.trip.presentation.trip.dto.response.TripPeriodUpdateResponse;
 import lombok.RequiredArgsConstructor;
