+++ src/test/java/com/cosain/trilo/unit/trip/application/trip/service/trip_image_update/TripImageUpdateServiceTest.java
@@ -2,7 +2,7 @@ package com.cosain.trilo.unit.trip.application.trip.service.trip_image_update;
 
 import com.cosain.trilo.common.file.ImageFile;
 import com.cosain.trilo.fixture.TripFixture;
-import com.cosain.trilo.trip.application.exception.NoTripUpdateAuthorityException;
+import com.cosain.trilo.common.exception.trip.NoTripUpdateAuthorityException;
 import com.cosain.trilo.common.exception.trip.TripNotFoundException;
 import com.cosain.trilo.trip.application.trip.service.trip_image_update.TripImageUpdateCommand;
 import com.cosain.trilo.trip.application.trip.service.trip_image_update.TripImageUpdateService;
