+++ src/test/java/com/cosain/trilo/unit/trip/application/trip/service/trip_period_update/TripPeriodUpdateCommandTest.java
@@ -2,8 +2,8 @@ package com.cosain.trilo.unit.trip.application.trip.service.trip_period_update;
 
 import com.cosain.trilo.common.exception.CustomValidationException;
 import com.cosain.trilo.trip.application.trip.service.trip_period_update.TripPeriodUpdateCommand;
-import com.cosain.trilo.trip.domain.exception.InvalidPeriodException;
-import com.cosain.trilo.trip.domain.exception.TooLongPeriodException;
+import com.cosain.trilo.common.exception.trip.InvalidPeriodException;
+import com.cosain.trilo.common.exception.trip.TooLongPeriodException;
 import com.cosain.trilo.trip.domain.vo.TripPeriod;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.DisplayName;
